seperate BrowserPage and RoomsList component
